Move coercer into data types module
