+++ spec/support/key_helper.rb
@@ -1,3 +1,3 @@
 def create_key(key_name)
-  Dynameister::Coercer.new.create_key(key_name)
+  Dynameister::DataTypes::Coercer.new.create_key(key_name)
 end
